DON-1207 Make extraneous / Staging-problematic DB migration run in 'regression' only
This should mean its added as "done" in other envs as a no-op, which is hopefully fine for Production because
(if we're luckier than on Staging), Version20260911123329 will cover the same.

+++ src/Migrations/Version20260911153909.php
@@ -19,6 +19,10 @@ final class Version20260911153909 extends AbstractMigration
 
     public function up(Schema $schema): void
     {
+        if (getenv('APP_ENV') !== 'regression') {
+            return;
+        }
+
         $this->addSql('DROP INDEX FULLTEXT_GLOBAL_SEARCH ON Campaign');
         $this->addSql('DROP INDEX FULLTEXT_NORMALISED_NAME ON Campaign');
         $this->addSql('DROP INDEX FULLTEXT_GLOBAL_SEARCH ON Charity');
